requirements.txt
pandas==2.1.2
pytest==7.4.3
python_dateutil==2.8.2
yfinance==0.2.31
